requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.1.0=pypi_0
antlr-python-runtime=4.7.2=py39h6e9494a_1003
appdirs=1.4.4=pyh9f0ad1d_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
brotli=1.0.9=h5eb16cf_7
brotli-bin=1.0.9=h5eb16cf_7
brotlipy=0.7.0=py39h63b48b0_1004
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=hca72f7f_0
ca-certificates=2022.4.26=hecd8cb5_0
cachetools=5.2.0=pypi_0
cartopy=0.20.2=py39hb53ae2a_3
certifi=2022.6.15=py39hecd8cb5_0
cf-units=3.0.1=py39hc89836e_2
cffi=1.15.0=py39hc55c11b_1
cftime=1.5.1.1=py39h67323c0_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
cloudpickle=2.0.0=pyhd8ed1ab_0
cryptography=37.0.2=py39h9c2a9ce_0
curl=7.80.0=hca72f7f_0
cycler=0.11.0=pyhd8ed1ab_0
dask-core=2022.4.0=pyhd8ed1ab_0
dill=0.3.4=pyhd3eb1b0_0
expat=2.4.4=he9d5cce_0
flatbuffers=1.12=pypi_0
fonttools=4.31.2=py39h63b48b0_1
freetype=2.10.4=h4cff582_1
fsspec=2022.3.0=pyhd8ed1ab_0
gast=0.4.0=pypi_0
geos=3.10.2=he49afe7_0
google-auth=2.8.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.46.3=pypi_0
h5py=3.7.0=pypi_0
hdf4=4.2.13=h39711bb_2
hdf5=1.10.6=hdbbcd12_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.4=py39h6e9494a_0
importlib_metadata=4.11.4=hd8ed1ab_0
importlib_resources=5.7.1=pyhd8ed1ab_1
intel-openmp=2021.4.0=hecd8cb5_3538
iris=3.2.1=pyhd8ed1ab_0
jbig=2.1=h0d85af4_2003
jinja2=3.1.1=pyhd8ed1ab_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h9ed2024_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=py39he9d5cce_0
krb5=1.19.2=hcd88c3b_0
lcms2=2.12=h577c468_0
lerc=2.2.1=h046ec9c_0
libbrotlicommon=1.0.9=h5eb16cf_7
libbrotlidec=1.0.9=h5eb16cf_7
libbrotlienc=1.0.9=h5eb16cf_7
libclang=14.0.1=pypi_0
libcurl=7.80.0=h6dfd666_0
libcxx=12.0.0=h2f01273_0
libdeflate=1.7=h35c211d_5
libedit=3.1.20210910=hca72f7f_0
libev=4.33=h9ed2024_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libmo_unpack=3.1.2=h0a44026_1001
libnetcdf=4.8.1=h24cb85c_1
libnghttp2=1.46.0=ha29bfda_0
libpng=1.6.37=h7cec526_2
libssh2=1.9.0=ha12b0ac_1
libtiff=4.3.0=h1167814_1
libwebp-base=1.2.2=h0d85af4_1
libzip=1.5.1=h08130ce_1003
llvm-openmp=12.0.0=h0dcd299_1
locket=0.2.0=py_2
lz4-c=1.9.3=he49afe7_1
markdown=3.3.7=pypi_0
markupsafe=2.1.1=py39h63b48b0_1
matplotlib-base=3.5.1=py39hb07454d_0
metpy=1.3.0=pyhd8ed1ab_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
mo_pack=0.2.0=py39hc89836e_1006
mogp-emulator=0.6.0=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=hca72f7f_2
netcdf4=1.5.7=py39h4a1dd59_1
numpy=1.21.2=py39h4b4dc7a_0
numpy-base=1.21.2=py39he0bd621_0
oauthlib=3.2.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1o=hca72f7f_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.1=py39h4d6be9b_0
partd=1.2.0=pyhd8ed1ab_0
patsy=0.5.2=pypi_0
pillow=7.2.0=py39h31b929e_2
pint=0.19.2=pyhd8ed1ab_0
pip=21.2.4=py39hecd8cb5_0
pooch=1.6.0=pyhd8ed1ab_0
progress=1.6=pypi_0
proj=8.2.1=h1512c50_0
protobuf=3.19.4=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyproj=3.3.0=py39h074cefc_1
pyshp=2.2.0=pyhd8ed1ab_0
pysocks=1.7.1=py39h6e9494a_5
python=3.9.12=hdfd78df_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-xxhash=3.0.0=py39h63b48b0_0
python_abi=3.9=2_cp39
pytz=2022.1=pyhd8ed1ab_0
pyyaml=6.0=py39h63b48b0_4
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=py39hae1ba45_1
scipy=1.7.3=py39h8c7af03_0
setuptools=58.0.4=py39hecd8cb5_0
shapely=1.8.0=py39hbfbc381_5
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=h707629a_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h7bc2e8c_0
toolz=0.11.2=pyhd8ed1ab_0
traitlets=5.2.1.post0=pyhd8ed1ab_0
typing_extensions=4.2.0=pyha770c72_1
tzdata=2022a=hda174b7_0
udunits2=2.2.28=h06ef574_0
unicodedata2=14.0.0=py39h63b48b0_1
urllib3=1.26.9=pyhd8ed1ab_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xarray=2022.3.0=pyhd8ed1ab_0
xxhash=0.8.0=h35c211d_3
xz=5.2.5=h1de35cc_0
yaml=0.2.5=h0d85af4_2
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.11=h4dc903c_4
zstd=1.5.0=hcb37349_1
